When will the testing end

+++ godot_steam_networking/addons/Networking/Helpers/SerializationExtension.cs
@@ -30,7 +30,7 @@ public static class SerializationExtension
             bytes.Add((byte)c);
         }
         return [
-            ..((ushort)bytes.Count).Serialize(),
+            ..bytes.Count.Serialize(),
             ..bytes.ToArray()
         ];
     }
